refactor evaluation sdk
