+++ client/starwhale/api/_impl/evaluation/pipeline.py
@@ -13,7 +13,7 @@ import dill
 import jsonlines
 
 from starwhale.utils import console, now_str
-from starwhale.consts import RunStatus, CURRENT_FNAME, DecoratorInjectAttr
+from starwhale.consts import RunStatus, CURRENT_FNAME
 from starwhale.utils.fs import ensure_dir, ensure_file
 from starwhale.api._impl import wrapper
 from starwhale.base.type import RunSubDirType, PredictLogMode
@@ -443,309 +443,3 @@ class PipelineHandler(metaclass=ABCMeta):
             **input_features,
         }
         self.evaluation_store.log_result(record)
-
-
-# TODO: add flush, get_summary functions?
-class EvaluationLogStore:
-    _instance_holder = threading.local()
-    _lock = threading.Lock()
-
-    def __init__(self, id: str, project: str) -> None:
-        self.id = id
-        self.project = project
-
-        self._datastore = wrapper.Evaluation(eval_id=self.id, project=self.project)
-
-    def __str__(self) -> str:
-        return f"Evaluation: id({self.id}), project({self.project})"
-
-    __repr__ = __str__
-
-    @classmethod
-    def _get_instance(cls) -> EvaluationLogStore:
-        with cls._lock:
-            _inst: t.Optional[EvaluationLogStore]
-            try:
-                _inst = cls._instance_holder.value  # type: ignore
-            except AttributeError:
-                _inst = None
-
-            if _inst is not None:
-                return _inst
-
-            context = Context.get_runtime_context()
-
-            _inst = cls(
-                id=context.version,
-                project=context.project,
-            )
-
-            cls._instance_holder.value = _inst
-            return _inst
-
-    @classmethod
-    def log(
-        cls, category: str, id: t.Union[str, int], metrics: t.Dict[str, t.Any]
-    ) -> None:
-        """Log metrics into the Starwhale Datastore.
-
-        Arguments:
-            category: [str, required] The category of the log records.
-            id: [Union[str, int], required] The unique id of the record.
-            metrics: [Dict, required] The metrics dict.
-
-        Examples:
-        ```python
-        from starwhale import evaluation
-
-        evaluation.log("label/1", 1, {"loss": 0.99, "accuracy": 0.98})
-        evaluation.log("ppl", "1", {"a": "test", "b": 1})
-
-        Returns:
-            None
-        ```
-        """
-        # TODO: support pickle serialize
-        el = cls._get_instance()
-        el._log(category, id, metrics)
-
-    def _log(
-        self, category: str, id: t.Union[str, int], metrics: t.Dict[str, t.Any]
-    ) -> None:
-        self._datastore.log(table_name=category, id=id, **metrics)
-
-    @classmethod
-    def log_summary(cls, *args: t.Any, **kw: t.Any) -> None:
-        """Log info into the Starwhale Datastore summary table.
-
-        Arguments:
-            *args: [dict, optional] use dicts to store summary info.
-            **kwargs: [optional] use kv to store summary info.
-
-        Examples:
-        ```python
-        from starwhale import evaluation
-
-        evaluation.log_summary(loss=0.99)
-        evaluation.log_summary(loss=0.99, accuracy=0.99)
-        evaluation.log_summary({"loss": 0.99, "accuracy": 0.99})
-        ```
-
-        Returns:
-            None
-        """
-        metrics = {}
-        if len(args) == 0:
-            metrics = kw
-        elif len(args) == 1:
-            if len(kw) > 0:
-                raise ParameterError(
-                    f"Args({args}) and kwargs({kw}) are specified at the same time"
-                )
-            if not isinstance(args[0], dict):
-                raise ParameterError(f"Args({args[0]}) is not dict type")
-
-            metrics = args[0]
-        else:
-            raise ParameterError(f"The number of args({args}) is greater than one")
-
-        metrics.update(kw)
-        el = cls._get_instance()
-        el._log_summary(metrics)
-
-    def _log_summary(self, metrics: t.Dict) -> None:
-        self._datastore.log_summary_metrics(metrics)
-
-    @classmethod
-    def iter(
-        cls,
-        category: str,
-    ) -> t.Iterator:
-        """Get an iterator for the log data of the specified category.
-
-        Arguments:
-            category: [str, required] The category of the log records.
-
-        Examples:
-        ```python
-        from starwhale import evaluation
-
-        results = [data for data in evaluation.iter("label/0")]
-        ```
-
-        Returns:
-            An iterator.
-        """
-        # TODO: support batch_size
-        el = cls._get_instance()
-        return el._iter(category)
-
-    def _iter(self, category: str) -> t.Iterator:
-        return self._datastore.get(category)
-
-
-def predict(*args: t.Any, **kw: t.Any) -> t.Any:
-    """Defines a predict function.
-
-    This function can be used as a decorator to define an evaluation-predict function that maps the dataset data into
-    the different predict workers.
-
-    Arguments:
-        resources: [Dict, optional] Resources for the predict task, such as cpu, memory, nvidia.com/gpu etc. Current only supports
-            the Server instance.
-        replicas: [int, optional] The number of the predict tasks. Default is 1.
-        batch_size: [int, optional] Number of samples per batch. Default is 1.
-        fail_on_error: [bool, optional] Fast fail on the exceptions in the predict function. Default is True.
-        auto_log: [bool, optional] Auto log the return values of the predict function and the according dataset rows. Default is True.
-        log_mode: [str, optional] When auto_log=True, the log_mode can be specified to control the log behavior. Options are `pickle` and `plain`. Default is `pickle`.
-        log_dataset_features: [List[str], optional] When auto_log=True, the log_dataset_features can be specified to control the log dataset features behavior.
-            Default is None, all dataset features will be logged. If the list is empty, no dataset features will be logged.
-        needs: [List[Callable], optional] The list of the functions that need to be executed before the predict function.
-
-    Examples:
-    ```python
-    from starwhale import evaluation
-
-    @evaluation.predict
-    def predict_image(data):
-        ...
-
-    @evaluation.predict(
-        dataset="mnist/version/latest",
-        batch_size=32,
-        replicas=4,
-        needs=[predict_image],
-    )
-    def predict_batch_images(batch_data)
-        ...
-
-    @evaluation.predict(
-        resources={"nvidia.com/gpu": 1,
-                "cpu": {"request": 1, "limit": 2},
-                "mem": 200 * 1024 * 1024},  # 200MB
-        log_mode="plain",
-    )
-    def predict_with_resources(data):
-        ...
-
-    @evaluation.predict(
-        replicas=1,
-        log_mode="plain",
-        log_dataset_features=["txt", "img", "label"],
-    )
-    def predict_with_selected_features(data):
-        ...
-    ```
-
-    Returns:
-        The decorated function.
-    """
-
-    # TODO: support runtime
-
-    if len(args) == 1 and len(kw) == 0 and callable(args[0]):
-        return predict()(args[0])
-    else:
-
-        def _wrap(func: t.Callable) -> t.Any:
-            _register_predict(func, **kw)
-            setattr(func, DecoratorInjectAttr.Predict, True)
-            return func
-
-        return _wrap
-
-
-def _register_predict(
-    func: t.Callable,
-    datasets: t.Optional[t.List[str]] = None,
-    resources: t.Optional[t.Dict[str, t.Any]] = None,
-    needs: t.Optional[t.List[t.Callable]] = None,
-    replicas: int = 1,
-    batch_size: int = 1,
-    fail_on_error: bool = True,
-    auto_log: bool = True,
-    log_mode: str = PredictLogMode.PICKLE.value,
-    log_dataset_features: t.Optional[t.List[str]] = None,
-) -> None:
-    from .job import Handler
-
-    Handler.register(
-        name="predict",
-        resources=resources,
-        needs=needs,
-        replicas=replicas,
-        require_dataset=True,
-        extra_kwargs=dict(
-            predict_batch_size=batch_size,
-            ignore_error=not fail_on_error,
-            predict_auto_log=auto_log,
-            predict_log_mode=log_mode,
-            predict_log_dataset_features=log_dataset_features,
-            dataset_uris=datasets,
-        ),
-        built_in=True,
-    )(func)
-
-
-def evaluate(*args: t.Any, **kw: t.Any) -> t.Any:
-    """Defines an evaluate function.
-
-    This function can be used as a decorator to define an evaluation-evaluate function that reduces the results of the
-    predict function.
-
-    Argument:
-        needs: [List[Callable], required] The list of the functions that need to be executed before the evaluate function.
-        use_predict_auto_log: [bool, optional] Passing the iterator of the predict auto-log results into the evaluate function.
-            Default is True.
-        resources: [Dict, optional] Resources for the predict task, such as memory, cpu, nvidia.com/gpu etc. Current only supports
-            the Server instance.
-
-    Examples:
-    ```python
-    from starwhale import evaluation
-
-    @evaluation.evaluate(needs=[predict_image])
-    def evaluate_results(predict_result_iter):
-        ...
-
-    @evaluation.evaluate(
-        use_predict_auto_log=False,
-        needs=[predict_image],
-    )
-    def evaluate_results():
-        ...
-    ```
-
-    Returns:
-        The decorated function.
-    """
-
-    def _wrap(func: t.Callable) -> t.Any:
-        _register_evaluate(func, **kw)
-        setattr(func, DecoratorInjectAttr.Evaluate, True)
-        return func
-
-    return _wrap
-
-
-def _register_evaluate(
-    func: t.Callable,
-    needs: t.Optional[t.List[t.Callable]] = None,
-    resources: t.Optional[t.Dict[str, t.Any]] = None,
-    use_predict_auto_log: bool = True,
-) -> None:
-    from .job import Handler
-
-    if not needs:
-        raise ValueError("needs is required for evaluate function")
-
-    Handler.register(
-        name="evaluate",
-        resources=resources,
-        replicas=1,
-        needs=needs,
-        extra_kwargs=dict(
-            predict_auto_log=use_predict_auto_log,
-        ),
-        built_in=True,
-    )(func)
